Plain structure for tests
