--- exporter/collector/default_collector_test.go
@@ -1,16 +1,0 @@
-package collector
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestMetricFactory(t *testing.T) {
-	SUT := NewDefaultCollector("test")
-	SUT.SetDescriptor("test_metric", "", nil)
-
-	metric := SUT.MakeGaugeMetric("test_metric", 1)
-
-	assert.Equal(t, SUT.GetDescriptor("test_metric"), metric.Desc())
-}
